magit-clone: Define as a transient prefix command
... and add suffixes for the basic variants.

By default `magit-clone' continues to act as before and a prefix
argument has to be used to make it act as a transient prefix command,
but the new option `magit-clone-always-transient' can be used to make
that unnecessary.

This breaks non-interactive uses of `magit-clone', but that command,
like most Magit commands, was never intended for non-interactive use.
`magit-clone-internal' can now be used non-interactively, which may
be desirable because that may also sets `remote.pushDefault', which
"git clone" itself doesn't do.  But as the name of that function
implies I recommend not doing that.